pageranks fixed

+++ pagerank.h
@@ -9,8 +9,8 @@
 #include "invertedIndex.h"
 
 void pageRankWeights(double damp, double diffPR, int maxIterations, Graph g);
-float pageRankIncoming(Graph g, int i);
-float Win(Graph g, int j, int i);
-float Wout(Graph g, int j, int i);
+double pageRankIncoming(Graph g, int i);
+double Win(Graph g, int j, int i);
+double Wout(Graph g, int j, int i);
 
 #endif
